+++ addons/imguitoolbar/imguitoolbar.h
@@ -286,8 +286,8 @@ public:
     }
 
     inline size_t getNumButtons() const {return buttons.size();}
-    inline Button* getButton(size_t i) {return (i < buttons.size())? &buttons[i] : NULL;}
-    inline const Button* getButton(size_t i) const {return (i < buttons.size())? &buttons[i] : NULL;}
+    inline Button* getButton(size_t i) {return ((int)i < buttons.size())? &buttons[i] : NULL;}
+    inline const Button* getButton(size_t i) const {return ((int)i < buttons.size())? &buttons[i] : NULL;}
     inline void addButton(const Button& button) {
         buttons.push_back(button);
     }
